+++ tests/test_settings.py
@@ -3078,7 +3078,7 @@ def main() -> None:
     defaults = {}
     setting_ids = set()
 
-    assert data.version == "3.5.3"
+    assert data.version == "3.5.4"
 
     gradient_name = localization["mod_name"]["en"]
     assert gradient_name.startswith("{#color(174,239,105)}B")